requirements.txt
altair==5.5.0
annotated-types==0.7.0
asn1crypto==1.5.1
asttokens==3.0.0
atpublic==5.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.13.1
blinker==1.4
cachetools==5.5.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
cloudpickle==2.2.1
colorama==0.4.4
comm==0.2.2
configobj==5.0.6
constantly==15.1.0
cryptography==44.0.0
debugpy==1.8.12
decorator==5.1.1
distro==1.7.0
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.16.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.43
greenlet==3.1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==8.10.0
narwhals==1.20.1
nest-asyncio==1.6.0
numpy==2.2.1
oauthlib==3.2.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.8.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyHamcrest==2.0.2
PyJWT==2.3.0
pyodbc==5.2.0
pyOpenSSL==24.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-magic==0.4.24
pytz==2022.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
requirements-parser==0.11.0
rich==13.9.4
SecretStorage==3.3.1
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
snowflake-cli==3.2.2
snowflake-connector-python==3.12.3
snowflake-snowpark-python==1.25.0
snowflake.core==0.12.1
sortedcontainers==2.4.0
soupsieve==2.6
speedtest-cli==2.1.3
SQLAlchemy==2.0.39
sqlparse==0.5.3
ssh-import-id==5.11
stack-data==0.6.3
streamlit==1.41.1
tenacity==9.0.0
teradatasql==20.0.0.21
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==22.1.0
typer==0.12.5
types-setuptools==75.8.0.20250110
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.5
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.13
zipp==1.0.0
zope.interface==5.4.0
